chore: Move helper functions into utils
